[APG-454] rollback changes

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/domain/entity/create/StaffEntity.kt
@@ -8,6 +8,7 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
+import java.math.BigInteger
 import java.util.UUID
 
 @Entity
@@ -19,7 +20,7 @@ data class StaffEntity(
   var id: UUID? = null,
 
   @Column(name = "staffId")
-  var staffId: Int? = null,
+  var staffId: BigInteger? = null,
 
   @Column(name = "first_name", nullable = false)
   var firstName: String,
