+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/restapi/model/StaffDetail.kt
@@ -2,9 +2,10 @@ package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.model
 
 import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.entity.create.AccountType
 import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.entity.create.PomType
+import java.math.BigInteger
 
 data class StaffDetail(
-  val staffId: Int,
+  val staffId: BigInteger,
   val firstName: String,
   val lastName: String,
   val primaryEmail: String,
